% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{annual_densities}
\alias{annual_densities}
\title{Calculate squirrel densities}
\usage{
annual_densities(con, ...)
}
\description{
These functions caluclate various density mesaurements for the KRSP grids.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
