% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-behaviour.r
\name{check_behaviour}
\alias{check_behaviour}
\alias{check_behaviour_loc}
\alias{check_behaviour_mode}
\alias{check_behaviour_time}
\title{Check records in behaviour table}
\usage{
check_behaviour(con, grid, year, observer)

check_behaviour_loc(con, grid, year, observer)

check_behaviour_mode(con, grid, year, observer)

check_behaviour_time(con, grid, year, observer)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; one or more grids to search. Defaults to all grids.}

\item{year}{integer; one or more years to search within. Defaults to all
years.}

\item{observer}{character; one or more observers to highlight errors for.
Defaults to all observers.}
}
\value{
A data frame of records that failed the checks.
}
\description{
Perform data integrity checks on the behaviour table. Run all checks at once
with \code{check_bahaviour()}, or perform individual checks with the specific
functions outlined below.
}
\section{Checks}{


The following checks have been implemented:

\itemize{
  \item \code{check_behaviour_loc}: all locs should conform to the standard
    format, e.g. A.1, 22.1, or -1.8.
  \item \code{check_behaviour_mode}: the \code{mode} of the observation, i.e.
    how it was taken, should never be empty.
  \item \code{check_behaviour_time}: in general, behaviour observations
    should all be taken between between 6am and 10pm, with the exception of
    nest locs, which may be night locs.
}
}

\examples{
con <- krsp_connect()
# run individual checks
check_behaviour_time(con, grid = "AG", year = 2013)
check_behaviour_loc(con, grid = c("AG", "JO"), observer = "MES")
check_behaviour_mode(con, year = 2010:2012)
# or perform them all at once
check_behaviour(con, grid = "JO", year = 2012)
}
