% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-collars.r
\name{check_collars}
\alias{check_collars}
\title{Check radio collars}
\usage{
check_collars(con, grid, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; one or more grids to search. Defaults to all grids.}

\item{year}{integer; year to search within. Defaults to current year.}
}
\value{
A data frame of records that failed the checks.
}
\description{
Check that radio collars have been correctly entered into the trapping table.
}
\section{Checks}{


The following checks have been implemented:

\enumerate{
  \item All trapping records with \code{radio} = 1 (new collar), 2 (collar
    on), 3 (collar change), or 4 (collar removed) should have a radio collar
    frequency in the \code{collar} field.
  \item Radio collar frequencies should all be 6 digits with no decimal. If a
    decimal is used the frequency will get rounded to the nearest integer,
    e.g. 150.231 -> 150.
  \item Any squirrel with \code{radio} = 2 (collar on), 3 (collar change) or
    4 (collar removed) should have a previous record with \code{radio} = 1
    (new collar).
  \item All squirrels that currently have no collar (\code{radio = 5}), but
    previously had a collar (\code{radio = 1-3}), should have record for a
    collar removal (\code{radio = 4}).
}
}

\examples{
con <- krsp_connect()
check_collars(con, grid = "KL", year = 2014)
}
