% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{code_mast_years}
\alias{code_mast_years}
\title{code mast years by year}
\usage{
`code_mast_years(.tbl, .gridCol, .yearCol)`
}
\arguments{
\item{.tbl}{A dataframe to calculate mast years}

\item{.gridCol}{a column to extract grid names from. Can accept standard and non-standard evaluated column names}

\item{.yearCol}{a column to extract years from. Can accept standard and non standard evaluated column names}
}
\description{
code mast years by year
}
\details{
The output of \code{code_mast_years()} will contain columns 'mast', 'Exp', and 'EXP_label'
Returns a \code{data.frame} or \code{tbl_df} object passed as an argument as \code{.tbl}
}
\examples{
\dontrun {
con <- krsp_connect(group = 'krsp-aws')
tbl(con, 'cones') \%>\% 
code_mast_years(Grid, Year)
} 
}
