% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone-summary.R
\name{cone_count}
\alias{cone_count}
\alias{annual_cone_counts}
\alias{cones_grids_years}
\title{calculate cone count summaries
These functions are helpers to calculate count counts on the krsp grids}
\usage{
cone_count(con, ...)

annual_cone_counts(con, years, ...)

cones_grids_years(con, grids, years, ...)
}
\arguments{
\item{con}{A \code{MySQLConnection} connection to the KRSP database}

\item{...}{Reserved for future extensions}

\item{years}{integer; one or more years to search within. Defaults to all years}

\item{grids}{character; one or more grids to search. Defaults to all grids.}
}
\value{
A \code{tbl_MySQLConnection} query that can be gathered into memory using \code{collect()}
}
\description{
\itemize{
\item \code{cone_count()} returns individual stake data
\item \code{annual_cone_counts()} computes cone counts by year
\item \code{grids_cones_years()} computes cone counts by grid and year
}
}
\examples{
\dontrun {
con <- krsp_connect(group = 'krsp-aws')
annual_cone_counts(con, years = 2010:2015)
cones_grids_years(con, grids = c("KL", "SU"))
cones_grids_years(con)
cones_grids_years(con, years = 2019, grids = "KL", "SU")
}
}
