% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-package.r
\docType{package}
\name{krsp}
\alias{krsp}
\alias{krsp-package}
\title{krsp: Working with the Kluane Red Squirrel Project Database}
\description{
This package is designed to help users bring data from the KRSP database into
R. In addition, common database queries are standardized as R functions to
ensure consistency. Many of the functions are wrappers around \code{dplyr} or
\code{RMySQL} functions. The vignettes demonstrates the use of this package:
\code{browseVignettes(package = "krsp")}
}
