% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-census-progress.r
\name{krsp_census_progress}
\alias{krsp_census_progress}
\title{Display a table of progress towards census}
\usage{
krsp_census_progress(con, grid, year, census)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; a single grid to map}

\item{year}{integer; year of census you're working on. Must be greater than
2012 when the new census table was implemented.}

\item{census}{character; are you completing the may or august census?}
}
\value{
A data frame of squirrel seen since the last census.
}
\description{
This function is to meant to be used to monitor progress towards completion
of the census. It lists all squirrels caught since the last census and notes
whether then have been entered into the current census.
}
\details{
The user must identify which census they are completing. If it is an August
census, then this function returns all squirrels caught between May 15 and
August 15. If it's a May census that is being completed, then this function
returns all squirrels seen up to May 15 of the given year.
}
\examples{
con <- krsp_connect()
krsp_census_progress(con, "JO", 2014, "may") \%>\%
  head()
}
