% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-censusmap.r
\name{krsp_censusmap}
\alias{krsp_censusmap}
\title{Display a map of middens comparing this and previous census}
\usage{
krsp_censusmap(con, grid, year, census, data)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; a single grid to map}

\item{year}{integer; year of census you're working on. Must be greater than
2012 when the new census table was implemented.}

\item{census}{character; are you completing the may or august census?}

\item{data}{logical; whether to just return the data instead of a plot.}
}
\value{
An interactive plot comparing censuses, or a data frame if
  \code{data == TRUE}.
}
\description{
This function is to meant to be used to monitor progress towards completion
of the census. It shows all middens from the previous census and whether or
not they have been entered in this census.
}
\details{
The user must identify which census they are completing. An August census
will be compared to the May census from that year, while a May census will be
compared to the previous year's August census.
}
\examples{
con <- krsp_connect()
krsp_censusmap(con, "JO", 2016, "may", data = TRUE) \%>\%
  head()
krsp_censusmap(con, "KL", 2016, "august")
}
