% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-collars.r
\name{krsp_collars}
\alias{krsp_collars}
\title{Radio collar status}
\usage{
krsp_collars(con, grid, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; one or more grids to search. Defaults to all grids.}

\item{year}{integer; year to search within. Defaults to current year.}
}
\value{
A data frame of all squirrels with collars on.
}
\description{
Generate a list of all squirrels that currently have collars on. A squirrel
is assumed to have a collar on if there is a record for a new radio collar (
`radio = 1`), but no matching radio collar off record (`radio = 4`).
}
\examples{
con <- krsp_connect()
krsp_collars(con, grid = "JO", year = 2014)
}
