% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-colours.r
\name{krsp_colours}
\alias{krsp_colours}
\title{Colour key}
\usage{
krsp_colours(con, grid, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; grid to display colour key for.}

\item{year}{integer; year to display colour key for. Defaults to current year.}
}
\value{
A data frame allowing identification of squirrels by colours. In this
  data frame there are three flags: \code{valid} indicates whether the
  colours are valid (allowing for at most 2 colours per ear),
  \code{duplicate} indicates whether that combination of colours is used on
  multiple squirrels, and \code{standard} indicates whether a squirrel has
  standard colouring, i.e. one colour in each ear and no mixing of wires,
  pipes, and discs.
}
\description{
Display a colour key based on the most recent trapping records for a given
grid and year.
}
\examples{
con <- krsp_connect()
krsp_colours(con, grid = "KL", year = 2014)
}
