% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-connect.r
\name{krsp_connect}
\alias{krsp_connect}
\title{Connect to the krsp database}
\usage{
krsp_connect(dbname = "krsp", host = "localhost", port = 0L,
  user = "root", password = "", group = NULL, ...)
}
\arguments{
\item{dbname}{database name; defaults to \code{krsp}}

\item{host, port}{character; host name and port of krsp database. Defaults to a local
instance of the database.}

\item{user, password}{character; username and password. For a local instance of the
database, these can typically be left as is; however, to connect to a remote
instance of the krsp database these will be user specific and must be
supplied.}

\item{group}{character; my.cnf option group}

\item{...}{Additional arguments passed on to \code{\link[dplyr]{src_mysql}}.}
}
\value{
A connection to the krsp database for use with \code{dplyr}.
}
\description{
\code{krsp_connect} is a wrapper for \code{\link[dplyr]{src_mysql}} that
connects to the KRSP database. By default, this function will connect to a
local copy of the database as root user with no password. When connecting to
a remote database, the host, username, and password must be supplied. This
information can be passed as parameters to \code{krsp_connect}, however, it's
much safer to store these authentication parameters in a my.cnf file. Consult
the vignette for this package for further details.
}
\examples{
con <- krsp_connect()
krsp_tables(con)
}
