% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-dashboard.r
\name{krsp_dashboard}
\alias{krsp_dashboard}
\title{Launch the Shiny Squirrels dashboard}
\usage{
krsp_dashboard(location, dbname = "krsp", host = "localhost",
  username = "root", password = "")
}
\arguments{
\item{location}{string; choose one of the predefined database connection
settings. "local" will connect to a local copy of the database as root user
with no password. "camp" is to be used at camp to connect to the database
on the head tech computer over the local network. If \code{location} is
set, then all other parameters are ignored.}

\item{dbname}{string; the name of the database to connect to, typically
"krsp".}

\item{host}{string; the IP address of the computer hosting the database.}

\item{username}{string; username for connection.}

\item{password}{string; password for connection.}
}
\description{
This Shiny app requires a connection to the KRSP database, which can be
accomplished in one of two ways. Either using using \code{location} to choose
one of the predefined connection settings or by setting the database name,
host, username, and password directly.
}
\examples{
\dontrun{
krsp_dashboard()
}
}
