% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-hitlist.r
\name{krsp_hitlist}
\alias{krsp_hitlist}
\title{List last year's breeders by parturition date}
\usage{
krsp_hitlist(con, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{year}{integer; year to generate hitlist for, e.g. if year = 2016,
breeders from 2015 will be listed.}
}
\value{
A tbl of all female squirrels from previous August's census,
   including location, in descending order of parturition date. Can be
   converted to a data.frame.
}
\description{
List all female squirrel from August census of previous year, including
location, in descending order of parturition date. This is meant to be a
to do list at the beginning of the season, i.e. the earliest breeders last
year should be caught first this year.
}
\examples{
con <- krsp_connect()
krsp_hitlist(con, 2016)
}
