% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-litter-lookup.r
\name{krsp_litter_lookup}
\alias{krsp_litter_lookup}
\title{Lookup up the juveniles in a litter}
\usage{
krsp_litter_lookup(con, year, squirrel_id, ln)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{year}{integer; year to search within. Defaults to current year.}

\item{squirrel_id}{integer; squirrel ID.}

\item{ln}{integer; litter number.}
}
\value{
A data frame of the juveniles in the given litter.
}
\description{
Lookup a litter based on year, squirrel ID, and litter number and return a
data frame of the juveniles in that litter.
}
\examples{
con <- krsp_connect()
krsp_litter_lookup(con, year = 2013, squirrel_id = 11059, ln = 1)
}
