% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-locmap.r
\name{krsp_locmap}
\alias{krsp_locmap}
\title{Map behaviour records on grid}
\usage{
krsp_locmap(con, grid, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; a single grid to map}

\item{year}{integer; defaults to current year}

\item{data}{logical; if TRUE return data frame instead of plotting}
}
\value{
Displays and returns a \code{ggvis} plot of rattle locations, unless
  \code{data} is TRUE, in which case a data frame is returned and nothing is
  plotted.
}
\description{
Create an interative plot of rattle or feeding behaviour on given grid and in
given year. In the plot, squirrels are colour coded, sex is distinguished by
different symbols, and further information is available in a tool tip on
mouseover. Plotting is done using \code{ggvis}, which creates an HTML object
that pops up in the Viewer pane in RStudio.
}
\examples{
con <- krsp_connect()
krsp_locmap(con, "JO", 2014, data = TRUE) \%>\%
  head
krsp_locmap(con, "KL", 2015)
}

