% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-pedigree.R
\name{krsp_pedigree}
\alias{krsp_pedigree}
\title{Calculate KRSP pedigree}
\usage{
krsp_pedigree(con, ...)
}
\arguments{
\item{con}{a \code{MySQLConnection} connection to the krsp database}

\item{...}{reserved for future extensions}
}
\description{
This function will create a pedigree from the flastall2 table in the cloud version of the krsp database.
The pedigree will be called 'krsp_pedigree'.  flastall2 contains all squirrels born in the study (including those not tagged).
Note this code untested
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
