% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-pool.r
\name{krsp_pool}
\alias{krsp_pool}
\title{Create krsp connection from a Pool}
\usage{
krsp_pool(pool)
}
\arguments{
\item{pool}{A \code{pool} object.}
}
\value{
A connection to the krsp database for use with \code{dplyr}.
}
\description{
The \code{pool} package is designed for managing database connectinos in
Shiny apps when multiple users may be creating connections to the database
simultaneously. This wrapper function for \code{\link[pool]{src_pool}}
returns a \code{krsp} database connection object from a pool.
}
\examples{
pool <- pool::dbPool(
  drv = RMySQL::MySQL(),
  dbname = "krsp",
  host = "localhost",
  username = "root",
  password = ""
)
krsp_pool(pool)
}
