% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-rattlemap.r
\name{krsp_rattlemap}
\alias{krsp_rattlemap}
\title{Map behaviour records on grid}
\usage{
krsp_rattlemap(con, grid, year, date_range, locx_range, locy_range, middens,
  data)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; a single grid to map}

\item{year}{integer}

\item{date_range}{vector of starting and ending dates as Date objects or
character representation of date in YMD format (e.g. "2016-05-25"); only
show records that are between these dates.}

\item{locx_range}{vector of length 2 giving upper and lower bounds of x locs,
either as character (e.g. \code{c("A", "H.5")}) or numeric (e.g.
\code{5.5, 10}).}

\item{locy_range}{vector of length 2 giving upper and lower bounds of y locs,
either as character (e.g. \code{c("5", "10.5")}) or numeric (e.g.
\code{5, 10.5}).}

\item{middens}{character; whether to show middens from the August ("august")
census, the May census ("may"), or not display middens at all ("none").}

\item{data}{logical; if TRUE return data frame instead of plotting}
}
\value{
Displays and returns a \code{ggvis} plot of rattle locations, unless
  \code{data} is TRUE, in which case a data frame is returned and nothing is
  plotted.
}
\description{
Create an interative plot of rattle or feeding behaviour on given grid and in
given year. In the plot, squirrels are colour coded, sex is distinguished by
different symbols, and further information is available in a tool tip on
mouseover. Plotting is done using \code{ggvis}, which creates an HTML object
that pops up in the Viewer pane in RStudio.
}
\details{
Records can be further filtered to a date range using the \code{date_range}
arguments, which is a vector of the from and to dates, respectively, both of
which must be in the same year. If \code{date_range} is provided is ignored.
}
\examples{
con <- krsp_connect()
krsp_rattlemap(con, "JO", 2014, data = TRUE) \%>\%
  head()
krsp_rattlemap(con, "KL", 2015)
# choose date range
krsp_rattlemap(con, "AG", date_range = c("2014-04-01", "2014-04-10"))
# choose loc range
krsp_rattlemap(con, "JO", 2010, locx_range = c("D", "H"),
  locy_range = c(5, 10))
}
