% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-sql.r
\name{krsp_sql}
\alias{krsp_sql}
\title{Run SQL query on database}
\usage{
krsp_sql(con, sql)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{sql}{SQL query to execute}
}
\value{
data.frame of query results
}
\description{
Run a pure SQL query on the KRSP database and return a data frame. To avoid
accidentally changing the database, only \code{SELECT} queries are permitted.
}
\examples{
con <- krsp_connect()
krsp_sql(con, "SELECT COUNT(*) AS n_squirrels FROM squirrel")
}
