% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-tables.r
\name{krsp_tables}
\alias{krsp_tables}
\title{List tables in the KRSP database}
\usage{
krsp_tables(con)
}
\arguments{
\item{con}{A connection to the KRSP database}
}
\value{
List of tables in KRSP database
}
\description{
List tables in the KRSP database
}
\examples{
con <- krsp_connect()
krsp_tables(con)
}
