% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-top.r
\name{krsp_top}
\alias{krsp_top}
\title{Top squirrelers}
\usage{
krsp_top(con, grid, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; a single study grid}

\item{year}{integer; year to search within. Defaults to all years.}
}
\value{
A data frame of squirrelers and number of trapping records, behaviour
  observations, and collars.
}
\description{
Generate a list of the top squirrelers based on a variety of metrics.
}
\examples{
con <- krsp_connect()
krsp_top(con, year = 2014) \%>\%
  head()
krsp_top(con) \%>\%
  head()
}

