% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc-to-numeric.r
\name{loc_to_numeric}
\alias{loc_to_numeric}
\title{Convert loc from alphanumeric to numeric}
\usage{
loc_to_numeric(x)
}
\arguments{
\item{x}{character; vector of alphanumeric grid locations.}
}
\value{
Numeric vector of grid locations.
}
\description{
Convert alphanumeric grid locations to purely numeric ones, e.g. A.1 becomes 1.1.
}
\examples{
loc_to_numeric(c("A.1", "T.5", "-1.1", "5.0", "AAA"))
}
