% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-connect.r
\name{krsp_connect}
\alias{krsp_connect}
\title{Connect to the krsp database}
\usage{
krsp_connect(dbname = "krsp", host = "localhost", port = 0L,
  user = "root", password = "", group = NULL, ...)
}
\arguments{
\item{dbname}{database name; defaults to \code{krsp}}

\item{host, port}{Host name and port of krsp database. Defaults to a local
instance of the database.}

\item{user, password}{Username and password. For a local instance of the
database, these can typically be left as; however, to connect to a remote
instance of the krsp database these will be user specific and must be
supplied.}

\item{...}{Additional arguments passed on to \code{\link[dplyr]{src_mysql}}.}
}
\value{
A character vector of table names
}
\description{
\code{krsp_connect} is a wrapper for \code{\link[dplyr]{src_mysql}} that
connects to the KRSP database. By default, this function will connect to a
local copy of the database. The host, username, and password must be supplied
if a remote database is to be used.
}
\examples{
con <- krsp_connect()
krsp_tables(con)
}

