% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-locmap.r
\name{krsp_locmap}
\alias{krsp_locmap}
\title{Map behaviour records on grid}
\usage{
krsp_locmap(con, grid, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; a single grid to map}

\item{year}{integer; defaults to current year}
}
\value{
Displays and returns a \code{ggvis} plot of rattle locations.
}
\description{
Create an interative plot of rattle or feeding behaviour on given grid and in
given year. In the plot, squirrels are colour coded, sex is distinguished by
different symbols, and further information is available in a tool tip on
mouseover. Plotting is done using \code{ggvis}, which creates an HTML object
that pops up in the Viewer pane in RStudio.
}
\examples{
con <- krsp_connect()
krsp_locmap(con, "AG", 2015)
}

