% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-needs-br.r
\name{krsp_needs_br}
\alias{krsp_needs_br}
\title{Identify squirrels without breeding status}
\usage{
krsp_needs_br(con, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{year}{integer; vector of years to search within. Defaults to current year.}
}
\value{
A tbl of squirrels, including their IDs and identifying
   information. Can be converted to a data.frame.
}
\description{
Identify female squirrels who are missing a breeding status (\code{br} in the
\code{litter} table) in the given year. Note that all squirrels should
appear in the litter table and have a breeding status even if they did not
breed in a given year.
}
\examples{
con <- krsp_connect()
krsp_needs_br(con, 2015)
}
