% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp-progress.r
\name{krsp_progress}
\alias{krsp_progress}
\title{Display seasonal data collection progress}
\usage{
krsp_progress(con, grid, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; a single grid to map}

\item{year}{integer; defaults to current year}

\item{data}{logical; if TRUE return data frame instead of plotting}
}
\value{
Displays and returns a \code{ggvis} plot of seasonal workflow
  progress for all females, unless \code{data} is TRUE, in which case a data
  frame is returned and nothing is plotted.
}
\description{
Display an interactive visualization of data collection progress for the
season. All females caught in the given year are inlcuded provided their most
recent trapping record has fate 1-3.
}
\examples{
con <- krsp_connect()
krsp_progress(con, "JO", 2015)
}

