% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krsp_progress.r
\name{krsp_progress}
\alias{krsp_progress}
\title{Display seasonal data collection progress}
\usage{
krsp_progress(con, grid, year)
}
\arguments{
\item{con}{Connection to KRSP database}

\item{grid}{character; a single grid to map}

\item{year}{integer; defaults to current year}
}
\value{
Displays and returns a \code{ggvis} plot of rattle locations.
}
\description{
Display seasonal data collection progress
}
\examples{
con <- krsp_connect()
krsp_progress(con, "AG", 2015)
}

